fix(Spehre3D): Moving intersection math to intersection3D module (pytest error)
pytest error collecting modules.
